afl-cmin.py: align with afl++'s new behavior'

+++ afl-cmin.py
@@ -46,14 +46,18 @@ group.add_argument('-o', dest='output', metavar='dir', required=True,
 group = parser.add_argument_group('Execution control settings')
 group.add_argument('-f', dest='stdin_file', metavar='file',
         help='location read by the fuzzed program (stdin)')
-group.add_argument('-m', dest='memory_limit', metavar='megs',
+group.add_argument('-m', dest='memory_limit', default='none', metavar='megs',
         type=lambda x: x if x == 'none' else int(x),
-        help='memory limit for child process (100)')
+        help='memory limit for child process (default: %(default)s)')
 group.add_argument('-t', dest='time_limit', default='none', metavar='msec',
         type=lambda x: x if x == 'none' else int(x),
-        help='timeout for each run (none)')
+        help='timeout for each run (default: %(default)s)')
+group.add_argument('-O', dest='frida_mode', action='store_true', default=False,
+        help='use binary-only instrumentation (FRIDA mode)')
 group.add_argument('-Q', dest='qemu_mode', action='store_true', default=False,
         help='use binary-only instrumentation (QEMU mode)')
+group.add_argument('-U', dest='unicorn_mode', action='store_true', default=False,
+        help='use unicorn-based instrumentation (Unicorn mode)')
 
 group = parser.add_argument_group('Minimization settings')
 group.add_argument('--crash-dir', dest='crash_dir', metavar='dir', default=None,
@@ -77,6 +81,7 @@ parser.add_argument('args', nargs='*')
 
 args = parser.parse_args()
 logger = None
+afl_showmap_bin = None
 
 def init():
     global logger
@@ -105,7 +110,7 @@ def init():
         logger.error('binary "%s" not found or not regular file', args.exe)
         sys.exit(1)
 
-    if not os.environ.get('AFL_SKIP_BIN_CHECK') and not args.qemu_mode:
+    if not os.environ.get('AFL_SKIP_BIN_CHECK') and not (args.qemu_mode or args.frida_mode or args.unicorn_mode):
         if b'__AFL_SHM_ID' not in open(args.exe, 'rb').read():
             logger.error("binary '%s' doesn't appear to be instrumented", args.exe)
             sys.exit(1)
@@ -114,6 +119,24 @@ def init():
         if not os.path.isdir(dn):
             logger.error('directory "%s" not found', dn)
             sys.exit(1)
+
+
+    global afl_showmap_bin
+    searches = [
+        None,
+        os.path.dirname(__file__),
+        os.getcwd(),
+        ]
+    if os.environ.get('AFL_PATH'):
+        searches.append(os.environ['AFL_PATH'])
+  
+    for search in searches:
+        afl_showmap_bin = shutil.which('afl-showmap', path=search)
+        if afl_showmap_bin:
+            break
+    if not afl_showmap_bin:
+        logger.fatal('cannot find afl-showmap, please set AFL_PATH')
+        sys.exit(1)
 
     trace_dir = os.path.join(args.output, '.traces')
     shutil.rmtree(trace_dir, ignore_errors=True)
@@ -128,18 +151,21 @@ def init():
         os.makedirs(args.crash_dir)
     os.makedirs(trace_dir)
 
-
 def afl_showmap(input_path, first=False):
     cmd = [
-            'afl-showmap',
+            afl_showmap_bin,
             '-m', str(args.memory_limit),
             '-t', str(args.time_limit),
             '-o', '-',
             '-Z']
     env = os.environ.copy()
 
+    if args.frida_mode:
+        cmd += ['-O']
     if args.qemu_mode:
         cmd += ['-Q']
+    if args.unicorn_mode:
+        cmd += ['-U']
     if args.edge_mode:
         cmd += ['-e']
     cmd += ['--', args.exe] + args.args
